Implementation Repeat

+++ instruction.cpp
@@ -2,14 +2,14 @@
 
 Instruction::Instruction() : m_id(IDs::Rien), m_parametres({}) {}
 
-Instruction::Instruction(IDs id, std::vector<int> parametres) : m_id(id), m_parametres(parametres) {}
+Instruction::Instruction(IDs id, std::vector<double> parametres) : m_id(id), m_parametres(parametres) {}
 
 void Instruction::setId(IDs id)
 {
     this->m_id = id;
 }
 
-void Instruction::setParametres(std::vector<int> parametres)
+void Instruction::setParametres(std::vector<double> parametres)
 {
     this->m_parametres = parametres;
 }
@@ -19,7 +19,7 @@ IDs Instruction::getId() const
     return this->m_id;
 }
 
-std::vector<int> Instruction::getParametres() const
+std::vector<double> Instruction::getParametres() const
 {
     return this->m_parametres;
 }
